Change locale tranlate query
This changes the util that generates the google translate formula so we
are complaint with google translate api specifications.

It uses ISO693-1. TLDR: It uses language identifiers and does not
distinguist language variants. `pt-br` works but is the same as
`pt-pt`, same happens for english, as this standard only cares about
the *mother language* (except a few cases that are handled).

More details: https://cloud.google.com/translate/docs/languages